Removed unused dependency
