+++ crates/dataflow-jit/Cargo.toml
@@ -20,7 +20,6 @@ lexical = "6.1.1"
 size-of = "0.1.4"
 tracing = "0.1.37"
 cranelift = "0.96.2"
-erased-serde = "0.3.25"
 enum_dispatch = "0.3.8"
 cranelift-jit = "0.96.2"
 target-lexicon = "0.12.5"
