+++ crates/dataflow-jit/src/codegen/vtable/clone.rs
@@ -28,7 +28,7 @@ impl Codegen {
         tracing::info!("creating clone vtable function for {layout_id}");
 
         // fn(*const u8, *mut u8)
-        let func_id = self.new_vtable_fn([self.module.isa().pointer_type(); 2], None);
+        let func_id = self.create_function([self.module.isa().pointer_type(); 2], None);
         let mut imports = self.intrinsics.import(self.comment_writer.clone());
 
         self.set_comment_writer(
@@ -122,7 +122,7 @@ impl Codegen {
 
         // fn(*const u8, *mut u8, usize)
         let ptr_ty = self.module.isa().pointer_type();
-        let func_id = self.new_vtable_fn([ptr_ty; 3], None);
+        let func_id = self.create_function([ptr_ty; 3], None);
         let mut imports = self.intrinsics.import(self.comment_writer.clone());
 
         self.set_comment_writer(
