+++ crates/dataflow-jit/src/codegen/vtable/hash.rs
@@ -14,7 +14,7 @@ impl Codegen {
         tracing::info!("creating hash vtable function for {layout_id}");
 
         // fn(&mut &mut dyn Hasher, *const u8)
-        let func_id = self.new_vtable_fn([self.module.isa().pointer_type(); 2], None);
+        let func_id = self.create_function([self.module.isa().pointer_type(); 2], None);
         let mut imports = self.intrinsics.import(self.comment_writer.clone());
 
         self.set_comment_writer(
